Rename controls2.tsx to controls.tsx
